requirements.txt
pandas==2.2.2
numpy==1.26.4
pyswisseph==2.10.3.2
python-dotenv==1.0.1
pytz==2024.1
seaborn==0.13.2
matplotlib==3.9.2
scipy==1.13.1

# pip install -r requirements.txt